requirements.txt
streamlit==1.37.1
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.40.0
ollama==0.3.2
email-validator==2.2.0
cryptography==43.0.1
langdetect==1.0.9
vaderSentiment==3.3.2
phonenumbers==8.13.45
geopy==2.4.1
